Added random pilot name generator
Method added to the Crew class that generates a random name of a gender based on the argument passed. This could be used when hiring new pilots.

Human name format: Firstname + Initial.
Robot name format: Alphabetical prefix + numerical suffix.

Added robot serial number prefix length and suffix length attributes, so we can vary the format of the serial number if we wish to. Perhaps different kinds of robots have different kinds of serial number.

Created buttons in the Crew panel for demo purposes.