Fixed listeners and enum names
Had a null reference exception for the Singleton.Instance. I think it was because the Awake() method of the Crew script was being called before the Awake() method of the Singleton (so the txt file hadn't been read in by the time the names field was being assigned in Crew). So I changed the Crew one to Start(). Seems to have fixed it but not sure if this is a good practice.