Changed end of round flow, fixed negative chip bug, and added method to deep copy the default deck (fixes index out of range bug)
Only the reference to the default deck was getting copied before, so the deck was being shuffled without having the drawn cards returned to it. This caused out of bounds error eventually.

Additional change: Now after each round, the player has to choose a new wager. This means that the chip check will occur, and the player won't be able to start a new game unless they can afford the wager.